fix process function
The change in the code is proposed because you should be able to pass an empty string and it process as null.  Plus the issue is some site showing how to use the function, for gulp for example, https://www.npmjs.com/package/gulp-vulcanize, show it as an empty string.  

The better way would be to trim the sides of whitespace so you could do  ' ' and it still return null.